chore(controller): split internal api components
